Data/requirements.txt
altgraph==0.17.2
cycler==0.11.0
Cython==0.29.26
docutils==0.18.1
fonttools==4.28.5
future==0.18.2
kivy-deps.angle==0.3.1
kivy-deps.glew==0.3.0
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.4
kivy-deps.sdl2-dev==0.4.4
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.5
pefile==2021.9.3
Pillow==9.0.0
Pygments==2.11.2
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
pyparsing==3.0.6
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
six==1.16.0
